Add tooltip for field conditions
(Hover over the turn marker.)

Megas made hazards difficult to see, and gigantamaxes are making
weather difficult to see. This has been a constant problem which can
be fixed through careful changing of the graphics, but we'd probably
need something like this sooner or later anyway.

This also refactors the tooltip placement code, which was previously
a bit messy.