Add reference to switch / fix typo
Hey, as I translated the this site to german, I noticed that solely the 3Ds is referenced when describing the RNG used in pokemon games. Now that the switch is the main platform for console pokemon games, it would be fitting to reference it aswell; also I fixed a typo provably->probably.