Refactor modded Pokedex data in teambuilder-tables, add mod elements to search support, add specific format mimicking
Modded pokedex entries now save any line that is different from the base dex, to avoid the number of specific named subtables from getting out of control.

Rudimentary search support was added, but needs to be upgraded because it isn't mod specific

Modded formats can set a "teambuilderFormat" and they will mimic that format for the purposes of the battle-dex-search file. This allows multiple formats from the same mod to work and allows tiers to show up properly in general. This could possibly be autodetected but I think the one line of code in config/formats.ts (on the server) is pretty reasonable