Update .babelrc
The new Babel version gives us access to transforming `||=` and `??`,
and removes the need for a lot of dependencies that are now included in
`@babel/preset-env` thanks to them being released rather than proposals.

This also adds a new Babel plugin to simply remove import/export when
building, which seems to be the best way to make `import` and `export`
usable on the client.