Start modularizing some battle files
Progress is very incomplete, but this is mostly an experiment to
establish this as possible at all.

What's going on here is that we're using `remove-import-export` so
the `import` and `export` statements just disappear after compiling
(everything's still a global).

This allows us to piecewise convert files to modules.

So we're just using TypeScript to keep track of dependencies, and
also to make it easier to use these files in other projects later on.

I've tried to avoid circular dependencies, but there's one between
`battle-animations` and `battle-animations-moves`.