Remove a reference to Lodash in replay-embed
(replay-embed is used to play back downloaded replay files)

In #907, we forgot to update a reference from `lodash.compat` to
`lodash.core`. Nothing broke, though, so this was never noticed.

lodash is no longer a dependency of most modern PS files (we've switched
to `ps-polyfill`, mostly). So it's better to just remove the dependency
here.

(Also add a header comment documenting the file a little.)