Desktop: Use Electron instead of NW.js
Electron makes the desktop client smaller and reduces the number of workarounds needed to account for differences between the desktop and browser clients. Packaging is also simplified, with the `electron-forge` tool replacing the multi-step NW.js process.

All the dependencies and code to be bundled for the desktop client are in the `desktop/` folder, with packaging instructions in `desktop/DESKTOP.md`. I have also created a GitHub Actions workflow to automatically rebuild the desktop client whenever it's changed, but we can remove that if it's not wanted.

macOS codesigning seems to be simplified by Electron's tooling, but I didn't test it myself because I didn't want to spend $100 on a developer account.