Website: Use an upscaled background
I heard some complaints about Horizon looking blurry on large screens, so
here I am spending nearly 600 KB of the home page's 1 MB bandwidth use on
a nice-looking background.

To be fair, it is in fact a nice-looking background and the most
eyecatching thing there, so it's probably worth it.