New Replays: Support private replays & more
The headline feature here is that we now support being logged in and
searching for private replays. We also have a new button to instantly
find your own replays.

Also here is the basics of HTML5 history support. It's the first step
to having the two-column layout up and running!