New Replays: Support linking to viewpoints
New Replays now supports adding `?p2` to the end of a replay
URL, to see the replay from the opposing viewpoint.

Search results will now automatically be from the viewpoint of
the player you're searching for.

Viewpoint is now a selector rather than a "Switch sides" button,
which should hopefully be clearer about what it's doing.