Refactor TypeScript files to modules
Due to the specific way we implement everything, none of these files
are actually modules. But thanks to `babel-plugin-remove-import-export`,
there's no reason why the code can't be written as modules.

I wouldn't say I like modules, but it's what's popular and it's what
people are used to and it's what tooling is built around.