Redesign website top navigation
The topbar in pokemonshowdown.com (and all not-in-sim pages)
has gotten revised in three ways:

1. Ladder has been replaced with Strategy, a link to the Smogon
   Strategydex.
2. Strategy and Forum now open in new tabs, because they don't
   share top nav with PS.
3. Strategy and Forum are now purple, to signal that they lead
   to Smogon and open in new tabs.

I'm way more a fan of the new look than I expected to be. Purple
and blue matches the "Horizon" art used on the main page quite
well, it turns out.

Ladder was always kind of useless - there's a better Ladder button
in-sim that keeps way more up-to-date with tier additions/removals,
and finding users can be done by clicking on the "X users online"
button.