Update intro comment
Unfortunately it seems like the console is the new hot place to say
hi to devs, rather than the source, but we're actually using the
console. So it's gotta be here.

ASCII art circle was outlined with

https://www.asciiart.eu/ascii-draw-studio

and the exclamation mark was done by tracing the old one. The rest,
including the anti-aliasing was entirely by hand, though.