Preact: Support reconnects
This adds support for reconnecting to the server after being
disconnected, without needing to refresh.

This is a little dangerous at the moment - one reason we've never done
this in the past is to enforce that everyone's using the newest client
version. This is fine in Preact Alpha but at some point we'll need to
find some way to enforce refreshing at a certain point.