Preact: Implement team copy/move
The clipboard can now handle teams as well as sets. The functionality
is pretty complex under the hood, but it should work basically as
expected.

If you copy sets into the team list, they'll become a new team
containing all the sets.

If you copy teams into the set list, all the sets inside the teams
will be added to the set list.