Consolidate into `teamSent`
In practice, it's impossible to send/accept a challenge at the same
time, so there's no need for two variables here.

In addition, "Checking..." isn't always accurate (it's not like randbat
has team validation), so I'm changing it to "Challenging..." and
"Accepting...". I've also made it not show a separate screen, so the
format/team selection isn't lost if validation fails.